Local Backup Working 27.11.23
Backup to local drives is working as intended, potentially sub folders to be added later on.

+++ main2.py
@@ -1,13 +1,19 @@
+import argparse
+import hashlib
 import shutil
 import os
 
-def backup_exported_photos(photo_folder, backup_location):
-    for root, dirs, files in os.walk(photo_folder):
+moved_file_count = 0
+
+def backup_exported_photos(photos, backup):
+    global moved_file_count
+
+    for root, dirs, files in os.walk(photos):
         for dir in dirs:
             if dir == "Export":
                 export_folder = os.path.join(root, dir)
                 parent_folder_name = os.path.basename(os.path.dirname(export_folder))
-                backup_folder = os.path.join(backup_location, parent_folder_name)
+                backup_folder = os.path.join(backup, parent_folder_name)
 
                 if not os.path.exists(backup_folder):
                     os.makedirs(backup_folder)
@@ -17,10 +23,37 @@ def backup_exported_photos(photo_folder, backup_location):
                         source_file = os.path.join(export_folder, filename)
                         destination_file = os.path.join(backup_folder, filename)
 
-                        shutil.copy(source_file, destination_file)
+                        if not os.path.exists(destination_file) or needs_update(source_file, destination_file):
+                            shutil.copy(source_file, destination_file)
+                            moved_file_count += 1
+
+def needs_update(source_file, destination_file):
+    source_hash = generate_md5_hash(source_file)
+    destination_hash = generate_md5_hash(destination_file) if os.path.exists(destination_file) else None
+
+    return source_hash != destination_hash
+
+
+def generate_md5_hash(filename):
+    with open(filename, "rb") as f:
+        hasher = hashlib.md5()
+
+        for chunk in iter(lambda: f.read(4096), b""):
+            hasher.update(chunk)
+
+        return hasher.hexdigest()
+
 
 if __name__ == "__main__":
-    photo_folder = "D:/BTest"
-    backup_location = "E:/After"
+    parser = argparse.ArgumentParser(description="Backup exported photos from a photo folder to a backup location")
+    parser.add_argument("--photodir", type=str, required=True, help="Path to the photo folder")
+    parser.add_argument("--backupdir", type=str, required=True, help="Path to the backup location")
 
-    backup_exported_photos(photo_folder, backup_location)
+    args = parser.parse_args()
+
+    photodir = args.photos
+    backupdir = args.backup
+
+    backup_exported_photos(photodir, backupdir)
+
+    print("Total moved or changed files:", moved_file_count)
